# src/main/java/javaspring/carservice/entity/Car.java
package javaspring.carservice.entity;

import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
@Builder
@Entity
@Table(name = "car")
@Data
@NoArgsConstructor
@AllArgsConstructor
public class Car {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @Column(nullable = false)
    private String model;

    @Column(nullable = false)
    private Integer enginePower;

    @Column(nullable = false)
    private Integer torque;

    @Column(name = "last_maintenance_timestamp")
    private LocalDateTime lastMaintenanceTimestamp;
}